Removed direct reference to media control object instead just referencing it from the cast_device object.
Reworked the cast device init stage to kill any
running and use a loop to wait for its idle
state.

Modified the update status call on the media
controller to use a count of 5 consecutive failed
attempts before declaring the device a write off.
Seems to better work around what seems to be a slow start
in getting the initial status