Reworked to self-load JSON config from ~/.castrc
The script can be invoked with --name or --ip and and optional --port
where you set the desired Chromecast detail by name or IP.

If these are omitted, then it sets itself up to monitor ~/.castrc
which contains a JSON-formatted config. This wil will be monitored
every 5 seconds and loaded if found to have changed.

The load will re-resolve the named Chromecast to IP and port
and the main agent state loop will detect a change in IP and port
and take the necessary actions to stop the older device and switch to
the new one