more tidy up
also fixed the chromecast -> volumio sync to only apply if the volumio
>= the chromecast elapsed time.. that will lessen the frequency of the
syncs but still let them make the check every 10 seconds.

Ideally they will occur once after a track is started and then once
after a volumio seek is performed, causing the cast playback to be
adjusted. The visual impact of the elapsed time niggles on the UI will be
lessened then.