rework of artwork model when using mpd. The mpd2 module is not giving back artwork and the suspicion is that even mpd itself is not doing this with the likes of Volumio and Moode doing that on their own.
So for this more abstract approach using an mpd module, I'm now trying to
use mutagen, a hadly python swiss army knife for accessing music file metadata to
extact the embedded files into /tmp and serving via URLs or falling back
to a search in the container directory of the playing file for suitable
images.

It feels seriously contrived to go about it this way as its deliberately
ignoring any existing Volumio/moOde solution but the approach is to try
and evolve this to a more abstract model of feeding off mpd only.