Pulled the initial mutagen impementation.. twas mad!
So now artwork is sourced from the native Volumio or moOde APIs.
At startup, an initial check for the volumio binary and similarly
moodeutl binary is made to to determine what we're running on. More
can be added to this check in time for other variants.

Then when casting the URL to the chromecast, a call is made to
get_artwork_url() that will check the determined variant and
access it's native API for one purpose only.. to get the URL
for the artwork.

It's the only reason we ever access the native APIs and the only
time we do it is when casting a new file or stream to the
chromecast. MPD API is used for all other playback activity.

Without that ability to call the local API, we'd have to retain
the mutagen approach and fallback to the local filesystem to determine
artwork.. an activity that both Volumio and moOde have already sorted.
So it makes no sense to have that duplication in play.