This script can be run against an MPD music directory where it will scan all directories for the presence of a cover.(png|jpg|tiff|bmp) file. If found, it skips the directory. If not found, it looks for the first flac, mp3. mp4 or m4a file it finds and tries to extract albumart from that file into a cover.xxx file.
It uses mutagen, the Python music file swiss army knife module to get
this done.

You can run this on a Pi as long as the given mount points for the music
is writable. If your mount it read-only, it will liely crash as it tries
to write the cover.xxx file