Changed music mount point to /var/lib/mpd/music as this is the MPD standard and how the releative pathis in fact managed. My bad for earlier thinking this was /mnt
Replaced albumart mechanism with simple trick of searching for
cover.(png|jpg|tiff|bmp) in the music file folder and serving that file
as the albumart URL.

Changed all naming of related functions and variables to albumart from
artwork

We are not (again) totally driving this via MPD and no longer using the
local APIs for albumart.

The mechanism of getting the albumart is a copy of what MPD itself does
internally. Alas the mpd2 client does not corectly support this and also
even if it did, it serves back the raw file, not a filename. I need the
filename as it's an URL we want to format, not serve the physical file
directly.