Added initial load of /tmp/castdevices into memory to allow for the web /cast page to have this available immediately after startup.
Put a try/except around the pychromecast scan as it seems to have issues
with socket resources after running for some time. Also set the returned
results (devices and browser) to None after we're done with the scan in
case there are some lingering background activities in the list of class
objects.

Changed handling of volume to remain in the MPD numbering of 0-100 and
only performs the chrome cast 0.0-1.0 conversion when it's passed to the
chromecast.

A few other changes in logging and how the chromecast wait is performed.
All part of a stability check for long-running playback.