A lot of updates to comments to bring them in line after such a lot of rework.
Some more tey/except rework here to batter manage unexpected API
scenarios.

Added a deadlock detection trick for the MPD agent. It sets a global
timestamp on each loop iteration to the curent time. If the main loop
detects 60 seconds gap on this timestamp from curent time, it exits.

The reason for this deadlock was more pychromecast-related than MPD.
For MPD, we call it's status on each loop. If that fails, we clear the
handle and reattempt a connection on the next run. It is pretty reliable
in mantaining the status and for reconection.

pychromecast however can get into a locked up state when it fails to
connect to a chromecast that has gone offline. We get our handle for
the chromecast from a call to pychromecast.get_chromecast_from_service().
That function has timeouts and retry limits but they do no seem to work
in that the function will not return after giving up.

So this deadlock then occurs and the main loop mechanism digs us out of
this