Reworked the positioning again of pause/play control to before the track cast as we ended up losing correct pause/resume behaviour. However I changed the unpause behavior to check that the cast track is the same as the curret one. This prevents a scenaio where when playing a track, we pause and then clange track.. what was happening was that we briefly unpaused the current cast track and then cast the new track.
Changed volume control to steps of 1 in either direction after improving
the smarts behind the scenes a bit.

Modified the page refresh of the /cast API to refrsh after three seconds
if the call had passed params. Otherwise it will default to 10 seconds.
This will get a better response to track and playback changes without
having to force a more frequent refresh all the time.