Removed bootstrap tooltips.. was unable to get them to work reliably. They kept getting stuck on screen despite multiple variations in trying to hide them on click and hover behaviours etc.
Changed playlist/queue refresh to be every 30 seconds as it was causing load
issues.

Changed the queue select combo to use the interna song ID as the action
value rather than a relative list index

Used the current playing song ID to auto-select the active song in the
queue.

Added support for Consume where the queue is gradualy purged of tracks
as they are played

Removed the onscreen volume indicator. Volume control is still a work in
progress. If changed too frequently, it locks up the MPD client due to
the agent thread (every second) and the web action code competing for
access.