removed the cast_file variable to get the logic to work completely using the cast_id and mpd_id matching.
For volumio, this is fine as MPD IDs always change for each "next" track
no matter what that is.. be it an actual next track or repeat of any
kind.

For moOde, its not as simple. moOde actually does play MPD more strictly
and the IDs will not change. So to work around this, I added tracking of
the chromecast native status.. PLAYING, IDLE etc and when it reads as
IDLE and the mpd_elapsed seconds is zero, it will recast the playing
track. This brings about desired repeat behaviour of a single repeat track or
list repeat where there is one track in the list