Updated platform detection to also determine albumart for a MPD stream with an attempt to detect Volumio or moOde default albumart images
Also modified the MPD stream cast to cast either a URL for the default
albumart or the platform title (moOde or Volumio).

Also modified the cherrypy code to host an extra static dir hosting of
the default albumart location. This was required to serve up the file to
the cast device.